pyproject.toml
[project]
name = "sentry-log-summarizer"
version = "0.1.0"
description = "Summarize Sentry event log attachments with a local LLM"
requires-python = ">=3.12"
dependencies = [
    "sentry-sdk>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "llm>=0.17.0",
    "llm-ollama>=0.5.0",
    "streamlit>=1.32.0",
    "watchdog>=4.0.0",
    "diskcache>=5.6.0",
    "chromadb>=0.6.0",
    "ddgs>=9.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "responses>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
